+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.5] - 20/03/2025
+
+### Added
+
+- Add support for `OpenAI`
+- Add support for `Anthropic`
+
 ## [0.1.4] - 18/03/2025
 
 ### Added